Split syntect into two crates
This allows editors to optionally use the theming and scope capacities
with custom parsers.

- closes #68
