+++ Cargo.toml
@@ -19,13 +19,13 @@ walkdir = "^1.0"
 regex-syntax = "^0.4"
 lazy_static = "^0.2"
 bitflags = "^0.8"
-plist = "0.2"
 bincode = "0.8"
 flate2 = "^0.2"
 fnv = "^1.0"
 serde = { version = "1.0", features = ["rc"] }
 serde_derive = "1.0"
 serde_json = "1.0"
+syntect-highlighting = { version = "1.3", path = "highlighting" }
 
 [dev-dependencies]
 regex = "0.2.1"
