+++ highlighting/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "syntect-highlighting"
+description = "scope and highlighting components for syntect"
+documentation = "http://thume.ca/rustdoc/syntect/syntect/"
+repository = "https://github.com/trishume/syntect"
+keywords = ["syntax", "highlighting", "highlighter", "colouring", "parsing"]
+readme = "Readme.md"
+license = "MIT"
+version = "1.3.0"
+authors = ["Tristan Hume <tristan@thume.ca>"]
+
+[dependencies]
+walkdir = "^1.0"
+lazy_static = "^0.2"
+bitflags = "^0.8"
+plist = "0.2"
+serde = { version = "1.0", features = ["rc"] }
+serde_derive = "1.0"
+serde_json = "1.0"
