+++ highlighting/src/scope.rs
