+++ highlighting/src/settings.rs
