+++ src/parsing/mod.rs
@@ -4,12 +4,10 @@ pub mod syntax_definition;
 #[cfg(feature = "yaml-load")]
 mod yaml_load;
 mod syntax_set;
-mod scope;
 mod parser;
 
 pub use self::syntax_definition::SyntaxDefinition;
 #[cfg(feature = "yaml-load")]
 pub use self::yaml_load::*;
 pub use self::syntax_set::*;
-pub use self::scope::*;
 pub use self::parser::*;
